+++ src/common/enum/possible-route-tests.js
@@ -3,9 +3,8 @@ function defineOptions()
 	var defineRes = {};
 	
 	defineRes["ONE_WAY"] = 1;
-	defineRes["SCENARIO_2"] = 2;
-	defineRes["SCENARIO_3"] = 3;
-	defineRes["SCENARIO_N"] = 4;
+	defineRes["START_SINGLE"] = 2;
+	defineRes["START_MULT"] = 3;
 	
 	return defineRes;
 }